Fraction models into separate files
