+++ app/models/application_record.rb
@@ -1,25 +1,3 @@
 class ApplicationRecord < ActiveRecord::Base
   primary_abstract_class
 end
-
-class User < ApplicationRecord
-  has_many :posts
-  has_many :likes
-  has_many :comments
-end
-
-class Post < ApplicationRecord
-  belongs_to :author, class_name: 'User', foreign_key: 'author_id'
-  has_many :likes
-  has_many :comments
-end
-
-class Like < ApplicationRecord
-  belongs_to :author, class_name: 'User', foreign_key: 'author_id'
-  belongs_to :post
-end
-
-class Comment < ApplicationRecord
-  belongs_to :author, class_name: 'User', foreign_key: 'author_id'
-  belongs_to :post
-end
