+++ app/models/comment.rb
@@ -1,0 +1,6 @@
+require_relative 'application_record'
+
+class Comment < ApplicationRecord
+  belongs_to :author, class_name: 'User', foreign_key: 'author_id'
+  belongs_to :post
+end
