+++ app/models/like.rb
@@ -1,0 +1,6 @@
+require_relative 'application_record'
+
+class Like < ApplicationRecord
+  belongs_to :author, class_name: 'User', foreign_key: 'author_id'
+  belongs_to :post
+end
