+++ app/models/post.rb
@@ -1,0 +1,7 @@
+require_relative 'application_record'
+
+class Post < ApplicationRecord
+  belongs_to :author, class_name: 'User', foreign_key: 'author_id'
+  has_many :likes
+  has_many :comments
+end
