+++ app/models/user.rb
@@ -1,0 +1,7 @@
+require_relative 'application_record'
+
+class User < ApplicationRecord
+  has_many :posts
+  has_many :likes
+  has_many :comments
+end
